# Speed up pytest execution wall time by ~37%.