@MatthiasJ1 2023-11-15
This is something we may want to open an issue for, I've noticed certain operations take an order of magnitude longer when inside a Builder context.