@fischman 2023-11-15
> This is something we may want to open an issue for, I've noticed certain operations take an order of magnitude longer when inside a Builder context.

Filed https://github.com/gumyr/build123d/issues/384 which also has a pointer to a self-contained repro test case, in case that's of interest.