@fischman 2023-11-15
> (full disclosure: the next-slowest test case is test_pack.py that I added very recently; I'll be looking into speeding that up next :) )

This now #386.